+++ providers/postgres/docs/changelog.rst
@@ -27,6 +27,17 @@
 Changelog
 ---------
 
+6.5.3
+.....
+
+Misc
+~~~~
+
+* ``Add SQLAlchemy to the dev‑dependency group in providers (#60472)``
+
+.. Below changes are excluded from the changelog. Move them to
+   appropriate section above if needed. Do not delete the lines(!):
+
 6.5.2
 .....